requirements_conda_env.txt
pip>=8.1.2
bumpversion>=0.5.3
wheel>=0.29.0
watchdog>=0.8.3
flake8>=2.6.0
tox>=2.3.1
coverage>=4.1
Sphinx>=1.4.8
cryptography>=1.7
PyYAML>=3.11
ruamel.yaml
pytest>=2.9.2
pytest-runner>=2.11.1
Click>=6.0
jupyter
jupyter_nbextensions_configurator
widgetsnbextension
jupyter_contrib_nbextensions
ipython
munch
pylama
logzero
pandas
numpy
appdirs
